[Refactor]#104 - extract ArtistSetlistCell
